with click event on markers to show infowindow, wikepedia
In this version, I'm trying to add the click event for the markers to display info window.  It seems like i need to have a closure, if I'm understanding it right, a function with a return value for the var marker.  Im going back to the lecture on that.... I feel like I have to redo my code.